правки лорданианского
With its roots in Pan-Slavic as well as in Zurich and Spacer, Lordanian became, primarily a language for the unification of the cultures of formerly independent West and East Lordania. The more popular Spacer began to wane after the rise of national unity and the desire to speak one's own language.
Shorthand: 'LRD'
It sounds like this:

CodexBot говорит, "Bian li kremverichda geskirana st. Skastach mykaen leskremich? Lesenis!"